[REFACTOR-024] i18n: split admin.ts into core/operations/content sub-domain files
