+++ lib/i18n/domains/admin.ts
@@ -1,417 +1,1 @@
-export const admin = {
-  'admin.nav.portal': { en: 'Portal', bg: 'Портал' },
-  'admin.nav.openMenu': { en: 'Open admin navigation', bg: 'Отвори навигацията' },
-
-  // -- Members Tab --
-  'admin.members.table.name': { en: 'Name', bg: 'Име' },
-  'admin.members.table.abo': { en: 'ABO', bg: 'СБА' },
-  'admin.members.table.sponsorAbo': { en: 'Sponsor ABO', bg: 'Номер на Спонсор' },
-  'admin.members.table.level': { en: 'Level', bg: 'Ниво' },
-  'admin.members.table.gpv': { en: 'GPV', bg: 'ГТС' },
-  'admin.members.table.bonus': { en: 'Bonus%', bg: 'Бонус%' },
-  'admin.members.table.group': { en: 'Group', bg: 'Група' },
-
-  'admin.members.btn.toCore': { en: '→ Core', bg: '→ Основен костяк' },
-  'admin.members.btn.toMember': { en: '→ Member', bg: '→ Член' },
-  'admin.members.noLosData': { en: 'No LOS data. Import a CSV in Data Center.', bg: 'Няма данни за LOS. Импортирайте CSV в Център за данни.' },
-  'admin.members.searchPlaceholder': { en: 'Search name or ABO…', bg: 'Търсене на ime или СБА…' },
-  'admin.members.btn.columns': { en: 'Columns', bg: 'Колони' },
-  'admin.members.pageInfo': { en: 'Page {{current}} of {{total}}', bg: 'Страница {{current}} от {{total}}' },
-  'admin.members.btn.prev': { en: 'Prev', bg: 'Предишна' },
-  'admin.members.btn.next': { en: 'Next', bg: 'Следваща' },
-
-  'admin.members.summary.losLinked': { en: '{{total}} in LOS · {{linked}} linked to portal accounts', bg: '{{total}} в LOS · {{linked}} свързани акаунти' },
-  'admin.members.pendingVerification': { en: 'Pending verification', bg: 'Чакащи потвърждение' },
-
-  'admin.members.verify.claimsAbo': { en: 'Claims ABO', bg: 'Заявява СБА' },
-  'admin.members.verify.upline': { en: 'upline', bg: 'горна линия' },
-  'admin.members.verify.los': { en: 'LOS:', bg: 'LOS:' },
-  'admin.members.verify.sponsorInLos': { en: 'sponsor in LOS:', bg: 'спонсор в LOS:' },
-  'admin.members.verify.losMatch': { en: '✓ LOS match', bg: '✓ LOS съвпадение' },
-  'admin.members.verify.uplineMismatch': { en: '⚠ upline mismatch', bg: '⚠ несъвпадение на горна линия' },
-  'admin.members.verify.noAboInLos': { en: '✗ ABO not in LOS', bg: '✗ СБА не е в LOS' },
-
-  'admin.members.verify.btn.approve': { en: 'Approve', bg: 'Одобри' },
-  'admin.members.verify.btn.deny': { en: 'Deny', bg: 'Откажи' },
-  'admin.members.verify.reasonPlaceholder': { en: 'Reason (optional)', bg: 'Причина (опционално)' },
-  'admin.members.verify.btn.confirm': { en: 'Confirm', bg: 'Потвърди' },
-
-  'admin.members.guestsNoAbo': { en: 'Guests — no ABO submitted ({{count}})', bg: 'Гости — без подаден СБА ({{count}})' },
-  'admin.members.guestJoined': { en: 'Joined', bg: 'Присъединен' },
-  'admin.members.guestRole': { en: 'guest', bg: 'гост' },
-  'admin.members.losMapDesc': { en: 'LOS map — {{count}} members', bg: 'LOS карта — {{count}} членове' },
-
-  // -- Data Center Tab --
-  'admin.data.title': { en: 'LOS CSV import — upserts on ABO number. Rebuilds LTree paths after every import.', bg: 'LOS CSV импорт — обновява по СБА. Преизгражда LTree пътищата след всеки импорт.' },
-  'admin.data.clickToSelect': { en: 'Click to select a CSV file', bg: 'Кликнете за избор на CSV файл' },
-  'admin.data.csvOnly': { en: '.csv only', bg: 'само .csv' },
-  'admin.data.rowsDetected': { en: '{{count}} rows detected', bg: '{{count}} открити реда' },
-  'admin.data.previewTitle': { en: 'Preview (first 5 rows)', bg: 'Преглед (първи 5 реда)' },
-  'admin.data.btn.runImport': { en: 'Run Import', bg: 'Стартирай импорт' },
-  'admin.data.btn.importing': { en: 'Importing...', bg: 'Импортиране...' },
-
-  'admin.data.result.complete': { en: 'Import complete — {{count}} rows upserted', bg: 'Импортът завърши — добавени/обновени {{count}} реда' },
-  'admin.data.result.new': { en: '{{count}} new', bg: '{{count}} нови' },
-  'admin.data.result.levelChanges': { en: '{{count}} level changes', bg: '{{count}} промени в нивото' },
-  'admin.data.result.bonusChanges': { en: '{{count}} bonus changes', bg: '{{count}} промени в бонуса' },
-  'admin.data.result.errors': { en: '{{count}} errors', bg: '{{count}} грешки' },
-  'admin.data.result.newMembersTitle': { en: 'New members', bg: 'Нови членове' },
-  'admin.data.result.rowErrorsTitle': { en: '{{count}} row errors — check for unsanitized data:', bg: '{{count}} грешки в редове — проверете за невалидни данни:' },
-  'admin.data.result.levelChangesTitle': { en: 'Level changes', bg: 'Промени в нивото' },
-  'admin.data.result.bonusChangesTitle': { en: 'Bonus % changes', bg: 'Промени в % бонус' },
-
-  'admin.data.reconciliation.title': { en: 'Reconciliation', bg: 'Свързване' },
-  'admin.data.reconciliation.desc': { en: 'Match unrecognized LOS members to manually-verified portal profiles.', bg: 'Свържете неразпознати LOS членове с ръчно верифицирани портал-профили.' },
-  'admin.data.reconciliation.btn.link': { en: 'Link', bg: 'Свържи' },
-  'admin.data.reconciliation.btn.linking': { en: 'Linking...', bg: 'Свързване...' },
-  'admin.data.reconciliation.unrecognizedTitle': { en: 'Unrecognized in LOS — {{count}}', bg: 'Неразпознати в LOS — {{count}}' },
-  'admin.data.reconciliation.allMatched': { en: 'All LOS members matched.', bg: 'Всички LOS членове са свързани.' },
-  'admin.data.reconciliation.noAboTitle': { en: 'No ABO — awaiting position — {{count}}', bg: 'Без СБА — очаква позиция — {{count}}' },
-  'admin.data.reconciliation.noManualMembers': { en: 'No manually-verified members without ABO.', bg: 'Няма ръчно верифицирани членове без СБА.' },
-  'admin.data.reconciliation.sponsor': { en: 'Sponsor', bg: 'Спонсор' },
-  'admin.data.reconciliation.upline': { en: 'Upline', bg: 'Горна линия' },
-
-  // -- LOS Tab --
-  'admin.los.treeDesc': { en: 'Full org tree with vital signs. Click a checkbox to toggle.', bg: 'Пълно организационно дърво с жизнени показатели. Кликнете, за да превключите.' },
-  'admin.los.memberCount': { en: '{{count}} members', bg: '{{count}} членове' },
-  'admin.los.vitalSigns': { en: 'Vital Signs', bg: 'Жизнени показатели' },
-  'admin.los.btn.active': { en: 'Active', bg: 'Активен' },
-  'admin.los.btn.inactive': { en: 'Inactive', bg: 'Неактивен' },
-  'admin.los.noAccountTooltip': { en: 'No portal account — cannot track vital signs', bg: 'Няма портал-акаунт — не могат да се проследят показатели' },
-  'admin.los.noData': { en: 'No LOS data yet.', bg: 'Все още няма LOS данни.' },
-
-  // -- Approval Hub --
-  'admin.approval.verify.standardTitle': { en: 'Standard ABO requests — {{count}}', bg: 'Стандартни СБА заявки — {{count}}' },
-  'admin.approval.verify.noStandard': { en: 'No pending standard requests.', bg: 'Няма чакащи стандартни заявки.' },
-  'admin.approval.verify.manualTitle': { en: 'Manual requests — {{count}}', bg: 'Ръчни заявки — {{count}}' },
-  'admin.approval.verify.noManual': { en: 'No pending manual requests.', bg: 'Няма чакащи ръчни заявки.' },
-
-  'admin.approval.verify.noAboBadge': { en: 'No ABO', bg: 'Без СБА' },
-  'admin.approval.verify.awaitingPosTitle': { en: 'Awaiting LOS positioning — {{count}}', bg: 'Очаква позициониране в LOS — {{count}}' },
-  'admin.approval.verify.approvedBadge': { en: 'approved', bg: 'одобрен' },
-
-  'admin.approval.verify.directTitle': { en: 'Direct verify', bg: 'Директна верификация' },
-  'admin.approval.verify.directDesc': { en: 'Directly promote a guest to member without a prior submission.', bg: 'Пряко повишаване на гост до член без предварителна заявка.' },
-  'admin.approval.verify.lbl.guest': { en: 'Guest', bg: 'Гост' },
-  'admin.approval.verify.lbl.uplineAbo': { en: 'Upline ABO number', bg: 'СБА номер на горна линия' },
-  'admin.approval.verify.opt.selectGuest': { en: 'Select a guest…', bg: 'Избери гост…' },
-  'admin.approval.verify.placeholder.upline': { en: 'e.g. 7010970187', bg: 'напр. 7010970187' },
-  'admin.approval.verify.btn.approve': { en: 'Approve', bg: 'Одобри' },
-  'admin.approval.verify.btn.deny': { en: 'Deny', bg: 'Откажи' },
-  'admin.approval.verify.btn.verifyMember': { en: 'Verify member', bg: 'Верифицирай член' },
-  'admin.approval.verify.btn.verifying': { en: 'Verifying…', bg: 'Верифициране…' },
-  'admin.approval.verify.btn.verified': { en: 'Verified ✓', bg: 'Верифициран ✓' },
-  'admin.approval.verify.noGuests': { en: 'No guests without a pending request.', bg: 'Няма гости без чакаща заявка.' },
-
-  'admin.approval.trips.resolvedCollapsible': { en: 'Resolved — {{count}}', bg: 'Разрешени — {{count}}' },
-  'admin.approval.trips.btn.allTrips': { en: 'All trips', bg: 'Всички пътувания' },
-  'admin.approval.trips.pendingTitle': { en: 'Pending — {{count}}', bg: 'Чакащи — {{count}}' },
-  'admin.approval.trips.noPending': { en: 'No pending registrations.', bg: 'Няма чакащи регистрации.' },
-
-  'admin.approval.events.btn.allEvents': { en: 'All events', bg: 'Всички събития' },
-  'admin.approval.events.noPending': { en: 'No pending role requests.', bg: 'Няма чакащи заявки за роли.' },
-
-  // -- Operations: Trips --
-  'admin.operations.trips.btn.new': { en: '+ New trip', bg: '+ Ново пътуване' },
-  'admin.operations.trips.empty': { en: 'No trips yet.', bg: 'Все още няма пътувания.' },
-  'admin.operations.trips.milestones': { en: '{{count}} milestones', bg: '{{count}} вноски' },
-  'admin.operations.trips.btn.edit': { en: 'Edit', bg: 'Редактирай' },
-  'admin.operations.trips.btn.delete': { en: 'Delete', bg: 'Изтрий' },
-  'admin.operations.trips.title.edit': { en: 'Edit: {{title}}', bg: 'Редактиране: {{title}}' },
-  'admin.operations.trips.dialog.title': { en: 'Delete trip', bg: 'Изтриване на пътуване' },
-  'admin.operations.trips.dialog.body': { en: 'This cannot be undone.', bg: 'Това действие е необратимо.' },
-  'admin.operations.trips.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.operations.trips.dialog.confirm': { en: 'Delete', bg: 'Изтрий' },
-
-  // -- Operations: TripForm --
-  'admin.operations.form.documents': { en: 'Trip documents', bg: 'Документи за пътуването' },
-  'admin.operations.form.loading': { en: 'Loading…', bg: 'Зареждане…' },
-  'admin.operations.form.noDocuments': { en: 'No documents uploaded yet.', bg: 'Все още няма качени документи.' },
-  'admin.operations.form.btn.uploading': { en: 'Uploading…', bg: 'Качване…' },
-  'admin.operations.form.btn.upload': { en: '+ Upload file', bg: '+ Качи файл' },
-  'admin.operations.form.uploadHint': { en: 'PDF or image, max 10 MB', bg: 'PDF или изображение, макс. 10 МБ' },
-  'admin.operations.form.btn.remove': { en: 'Remove', bg: 'Премахни' },
-  'admin.operations.form.dialog.title': { en: 'Remove document', bg: 'Премахване на документ' },
-  'admin.operations.form.dialog.removeConfirm': { en: 'Remove "{{name}}"?', bg: 'Премахване на "{{name}}"?' },
-  'admin.operations.form.dialog.body': { en: 'This cannot be undone.', bg: 'Това действие е необратимо.' },
-  'admin.operations.form.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.operations.form.dialog.confirm': { en: 'Remove', bg: 'Премахни' },
-  'admin.operations.form.lbl.description': { en: 'Description', bg: 'Описание' },
-  'admin.operations.form.lbl.startDate': { en: 'Start date', bg: 'Начална дата' },
-  'admin.operations.form.lbl.endDate': { en: 'End date', bg: 'Крайна дата' },
-  'admin.operations.form.lbl.cost': { en: 'Cost (EUR)', bg: 'Цена (EUR)' },
-  'admin.operations.form.lbl.milestones': { en: 'Payment milestones', bg: 'Вноски' },
-  'admin.operations.form.placeholder.label': { en: 'Label', bg: 'Наименование' },
-  'admin.operations.form.placeholder.amount': { en: 'Amount', bg: 'Сума' },
-  'admin.operations.form.btn.addMilestone': { en: '+ Add', bg: '+ Добави' },
-  'admin.operations.form.lbl.location': { en: 'Location', bg: 'Местоположение' },
-  'admin.operations.form.lbl.tripType': { en: 'Trip type', bg: 'Тип пътуване' },
-  'admin.operations.form.lbl.accommodation': { en: 'Accommodation', bg: 'Настаняване' },
-  'admin.operations.form.lbl.inclusions': { en: 'Inclusions (comma-sep)', bg: 'Включено (разделено със запетая)' },
-  'admin.operations.form.lbl.visibleTo': { en: 'Visible to', bg: 'Видимо за' },
-  'admin.operations.form.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
-  'admin.operations.form.btn.saveChanges': { en: 'Save changes', bg: 'Запази промените' },
-  'admin.operations.form.btn.createTrip': { en: 'Create trip', bg: 'Създай пътуване' },
-  'admin.operations.form.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
-
-  // -- Operations: Payments --
-  'admin.operations.payments.btn.log': { en: 'Log Payment', bg: 'Регистрирай плащане' },
-  'admin.operations.payments.pendingTitle': { en: 'Pending submissions ({{count}})', bg: 'Чакащи подавания ({{count}})' },
-  'admin.operations.payments.viewProof': { en: 'View proof ↗', bg: 'Виж доказателство ↗' },
-  'admin.operations.payments.badge.pending': { en: 'pending', bg: 'чакащо' },
-  'admin.operations.payments.filter.all': { en: 'All', bg: 'Всички' },
-  'admin.operations.payments.filter.pending': { en: 'Pending', bg: 'Чакащи' },
-  'admin.operations.payments.filter.approved': { en: 'Approved', bg: 'Одобрени' },
-  'admin.operations.payments.filter.rejected': { en: 'Rejected', bg: 'Отказани' },
-  'admin.operations.payments.empty': { en: 'No payments found.', bg: 'Няма намерени плащания.' },
-  'admin.operations.payments.drawer.title': { en: 'Log Payment', bg: 'Регистрирай плащане' },
-  'admin.operations.payments.lbl.entity': { en: 'Entity *', bg: 'Обект *' },
-  'admin.operations.payments.placeholder.entity': { en: 'Select entity…', bg: 'Избери обект…' },
-  'admin.operations.payments.group.trips': { en: 'Trips', bg: 'Пътувания' },
-  'admin.operations.payments.group.items': { en: 'Items', bg: 'Артикули' },
-  'admin.operations.payments.lbl.member': { en: 'Member *', bg: 'Член *' },
-  'admin.operations.payments.placeholder.member': { en: 'Select member…', bg: 'Избери член…' },
-  'admin.operations.payments.lbl.amount': { en: 'Amount *', bg: 'Сума *' },
-  'admin.operations.payments.lbl.currency': { en: 'Currency', bg: 'Валута' },
-  'admin.operations.payments.lbl.date': { en: 'Date', bg: 'Дата' },
-  'admin.operations.payments.lbl.method': { en: 'Payment method', bg: 'Начин на плащане' },
-  'admin.operations.payments.placeholder.method': { en: 'e.g. bank transfer, cash', bg: 'напр. банков превод, в брой' },
-  'admin.operations.payments.lbl.status': { en: 'Status', bg: 'Статус' },
-  'admin.operations.payments.lbl.note': { en: 'Note', bg: 'Бележка' },
-  'admin.operations.payments.lbl.noteOptional': { en: '(optional)', bg: '(незадължително)' },
-  'admin.operations.payments.placeholder.note': { en: 'e.g. Cash payment, ref #123', bg: 'напр. Плащане в брой, реф. #123' },
-  'admin.operations.payments.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
-  'admin.operations.payments.btn.log2': { en: 'Log payment', bg: 'Регистрирай плащане' },
-  'admin.operations.payments.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.operations.payments.btn.approve': { en: 'Approve', bg: 'Одобри' },
-  'admin.operations.payments.btn.deny': { en: 'Deny', bg: 'Откажи' },
-  'admin.operations.payments.placeholder.adminNote': { en: 'Admin note (optional)', bg: 'Бележка на администратор (незадължително)' },
-  'admin.operations.payments.status.approved': { en: 'Approved', bg: 'Одобрено' },
-  'admin.operations.payments.status.pending': { en: 'Pending', bg: 'Чакащо' },
-
-  // -- Operations: Items --
-  'admin.operations.items.btn.new': { en: '+ New item', bg: '+ Нов артикул' },
-  'admin.operations.items.empty': { en: 'No payable items yet.', bg: 'Все още няма артикули.' },
-  'admin.operations.items.btn.edit': { en: 'Edit', bg: 'Редактирай' },
-  'admin.operations.items.btn.deactivate': { en: 'Deactivate', bg: 'Деактивирай' },
-  'admin.operations.items.badge.active': { en: 'active', bg: 'активен' },
-  'admin.operations.items.badge.inactive': { en: 'inactive', bg: 'неактивен' },
-  'admin.operations.items.title.edit': { en: 'Edit: {{title}}', bg: 'Редактиране: {{title}}' },
-  'admin.operations.items.lbl.title': { en: 'Title *', bg: 'Наименование *' },
-  'admin.operations.items.lbl.type': { en: 'Type *', bg: 'Тип *' },
-  'admin.operations.items.lbl.description': { en: 'Description', bg: 'Описание' },
-  'admin.operations.items.lbl.amount': { en: 'Amount *', bg: 'Сума *' },
-  'admin.operations.items.lbl.currency': { en: 'Currency', bg: 'Валута' },
-  'admin.operations.items.toggle.active': { en: 'Active', bg: 'Активен' },
-  'admin.operations.items.toggle.inactive': { en: 'Inactive', bg: 'Неактивен' },
-  'admin.operations.items.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
-  'admin.operations.items.btn.saveChanges': { en: 'Save changes', bg: 'Запази промените' },
-  'admin.operations.items.btn.create': { en: 'Create item', bg: 'Създай артикул' },
-  'admin.operations.items.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.operations.items.type.merchandise': { en: 'merchandise', bg: 'стока' },
-  'admin.operations.items.type.ticket': { en: 'ticket', bg: 'билет' },
-  'admin.operations.items.type.food': { en: 'food', bg: 'храна' },
-  'admin.operations.items.type.book': { en: 'book', bg: 'книга' },
-  'admin.operations.items.type.other': { en: 'other', bg: 'друго' },
-
-  // -- Notifications audit log --
-  'admin.notifications.pageTitle': { en: 'Notification Audit Log', bg: 'Одиторски журнал на известията' },
-  'admin.notifications.pageDesc': { en: 'All system notifications ever fired — including soft-deleted. Read-only.', bg: 'Всички системни известия — включително изтритите. Само за четене.' },
-  'admin.notifications.errorLoading': { en: 'Error loading notifications:', bg: 'Грешка при зареждане на известия:' },
-  'admin.notifications.col.created': { en: 'Created', bg: 'Създадено' },
-  'admin.notifications.col.type': { en: 'Type', bg: 'Тип' },
-  'admin.notifications.col.title': { en: 'Title', bg: 'Заглавие' },
-  'admin.notifications.col.member': { en: 'Member', bg: 'Член' },
-  'admin.notifications.col.read': { en: 'Read', bg: 'Прочетено' },
-  'admin.notifications.col.deleted': { en: 'Deleted', bg: 'Изтрито' },
-  'admin.notifications.badge.read': { en: 'Read', bg: 'Прочетено' },
-  'admin.notifications.badge.unread': { en: 'Unread', bg: 'Непрочетено' },
-  'admin.notifications.empty': { en: 'No notifications found.', bg: 'Няма намерени известия.' },
-  'admin.notifications.pagination.info': { en: 'Page {{page}} of {{total}} — {{count}} total', bg: 'Страница {{page}} от {{total}} — {{count}} общо' },
-  'admin.notifications.pagination.prev': { en: '← Prev', bg: '← Предишна' },
-  'admin.notifications.pagination.next': { en: 'Next →', bg: 'Следваща →' },
-
-  // -- Settings --
-  'admin.settings.pageTitle': { en: 'System Settings', bg: 'Системни настройки' },
-  'admin.settings.pageDesc': { en: 'Manage global configurations and monitor automated systems', bg: 'Управлявайте глобалните конфигурации и наблюдавайте автоматизираните системи' },
-
-  // -- Settings: Email Log --
-  'admin.settings.emailLog.title': { en: 'Email Log', bg: 'Журнал на имейлите' },
-  'admin.settings.emailLog.desc': { en: 'Every outbound email attempt is recorded here. Use Retry on failed rows.', bg: 'Всеки опит за изпращане на имейл се записва тук. Използвайте Повторен опит за неуспешни редове.' },
-  'admin.settings.emailLog.col.template': { en: 'Template', bg: 'Шаблон' },
-  'admin.settings.emailLog.col.recipient': { en: 'Recipient', bg: 'Получател' },
-  'admin.settings.emailLog.col.sentAt': { en: 'Sent at', bg: 'Изпратено в' },
-  'admin.settings.emailLog.col.status': { en: 'Status', bg: 'Статус' },
-  'admin.settings.emailLog.filter.all': { en: 'All', bg: 'Всички' },
-  'admin.settings.emailLog.filter.sent': { en: 'sent', bg: 'изпратени' },
-  'admin.settings.emailLog.filter.failed': { en: 'failed', bg: 'неуспешни' },
-  'admin.settings.emailLog.filter.pending': { en: 'pending', bg: 'чакащи' },
-  'admin.settings.emailLog.allTemplates': { en: 'All templates', bg: 'Всички шаблони' },
-  'admin.settings.emailLog.empty': { en: 'No email log entries found.', bg: 'Няма намерени записи в журнала.' },
-  'admin.settings.emailLog.retryError': { en: 'Retry error:', bg: 'Грешка при повторен опит:' },
-  'admin.settings.emailLog.btn.retry': { en: 'Retry', bg: 'Повторен опит' },
-  'admin.settings.emailLog.btn.retrying': { en: 'Retrying…', bg: 'Повторен опит…' },
-  'admin.settings.emailLog.pagination.prev': { en: '← Prev', bg: '← Предишна' },
-  'admin.settings.emailLog.pagination.next': { en: 'Next →', bg: 'Следваща →' },
-
-  // -- Settings: Email Settings Panel --
-  'admin.settings.emailPanel.globalTitle': { en: 'System-wide Email', bg: 'Системни имейли' },
-  'admin.settings.emailPanel.globalDesc': { en: 'Enable or disable all automated outbound emails.', bg: 'Включете или изключете всички автоматизирани изходящи имейли.' },
-  'admin.settings.emailPanel.recipientTitle': { en: 'Admin Alert Recipient', bg: 'Получател на системни сигнали' },
-  'admin.settings.emailPanel.recipientDesc': { en: 'Receives bcc or system alerts if configured.', bg: 'Получава скрито копие или системни сигнали при конфигурация.' },
-  'admin.settings.emailPanel.btn.save': { en: 'Save', bg: 'Запази' },
-  'admin.settings.emailPanel.automatedTitle': { en: 'Automated Notifications', bg: 'Автоматизирани известия' },
-  'admin.settings.emailPanel.template.welcome': { en: 'Welcome Email', bg: 'Приветствен имейл' },
-  'admin.settings.emailPanel.template.payment_status': { en: 'Payment Status Updates', bg: 'Обновяване на статус на плащане' },
-  'admin.settings.emailPanel.template.doc_expiry': { en: 'Document Expiry Warnings', bg: 'Предупреждения за изтичащи документи' },
-  'admin.settings.emailPanel.template.abo_verification_result': { en: 'ABO Verification Results', bg: 'Резултати от СБА верификация' },
-  'admin.settings.emailPanel.template.trip_registration_status': { en: 'Trip Registration Updates', bg: 'Обновяване на регистрация за пътуване' },
-
-  // -- Content: page --
-  'admin.content.page.heading': { en: 'Content', bg: 'Съдържание' },
-  'admin.content.page.tab.news':    { en: 'News',         bg: 'Новини'        },
-  'admin.content.page.tab.links':   { en: 'Links',        bg: 'Връзки'        },
-  'admin.content.page.tab.guides':  { en: 'Guides',       bg: 'Ръководства'   },
-  'admin.content.page.tab.socials': { en: 'Social Posts', bg: 'Публикации'    },
-
-  // -- Content: Social --
-  'admin.content.social.btn.addPost': { en: '+ Add post', bg: '+ Добави публикация' },
-  'admin.content.social.drawer.newTitle': { en: 'New post', bg: 'Нова публикация' },
-  'admin.content.social.drawer.editTitle': { en: 'Edit: {{name}}', bg: 'Редактиране: {{name}}' },
-  'admin.content.social.placeholder.postUrl': { en: 'Post URL (required)', bg: 'URL на публикацията (задължително)' },
-  'admin.content.social.fetchingPreview': { en: 'Fetching preview…', bg: 'Зареждане на преглед…' },
-  'admin.content.social.previewUnavailable': { en: 'Preview unavailable for this platform — enter thumbnail URL manually', bg: 'Прегледът не е наличен за тази платформа — въведете URL на миниатюрата ръчно' },
-  'admin.content.social.placeholder.caption': { en: 'Caption — auto-extracted from post if left blank', bg: 'Надпис — автоматично извлечен от публикацията' },
-  'admin.content.social.placeholder.thumbnailUrl': { en: 'Thumbnail URL — auto-extracted from post if left blank', bg: 'URL на миниатюра — автоматично извлечен от публикацията' },
-  'admin.content.social.btn.uploading': { en: 'Uploading…', bg: 'Качване…' },
-  'admin.content.social.btn.uploadThumbnail': { en: 'Upload thumbnail', bg: 'Качи миниатюра' },
-  'admin.content.social.thumbnailStored': { en: 'Stored', bg: 'Съхранено' },
-  'admin.content.social.cdnUrlWarning': { en: 'This URL is an ephemeral CDN link and will expire. Upload a file instead.', bg: 'Този URL е временна CDN връзка и ще изтече. Качете файл вместо това.' },
-  'admin.content.social.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
-  'admin.content.social.btn.save': { en: 'Save', bg: 'Запази' },
-  'admin.content.social.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.content.social.badge.pinned': { en: 'Pinned', bg: 'Закачено' },
-  'admin.content.social.badge.active': { en: 'Active', bg: 'Активно' },
-  'admin.content.social.badge.hidden': { en: 'Hidden', bg: 'Скрито' },
-  'admin.content.social.btn.unpin': { en: 'Unpin', bg: 'Откачи' },
-  'admin.content.social.btn.pin': { en: 'Pin', bg: 'Закачи' },
-  'admin.content.social.btn.edit': { en: 'Edit', bg: 'Редактирай' },
-  'admin.content.social.btn.delete': { en: 'Delete', bg: 'Изтрий' },
-  'admin.content.social.empty': { en: 'No posts yet.', bg: 'Все още няма публикации.' },
-  'admin.content.social.dialog.title': { en: 'Delete post', bg: 'Изтриване на публикация' },
-  'admin.content.social.dialog.body': { en: 'Delete this post? This cannot be undone.', bg: 'Изтриване на тази публикация? Действието е необратимо.' },
-  'admin.content.social.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.content.social.dialog.confirm': { en: 'Delete', bg: 'Изтрий' },
-
-  // -- Content: Guides --
-  'admin.content.guides.btn.new': { en: '+ New Guide', bg: '+ Ново ръководство' },
-  'admin.content.guides.drawer.newTitle': { en: 'New Guide', bg: 'Ново ръководство' },
-  'admin.content.guides.drawer.editTitle': { en: 'Edit: {{name}}', bg: 'Редактиране: {{name}}' },
-  'admin.content.guides.empty': { en: 'No guides yet. Create the first one.', bg: 'Все още няма ръководства. Създайте първото.' },
-  'admin.content.guides.btn.publish': { en: 'Publish', bg: 'Публикувай' },
-  'admin.content.guides.btn.unpublish': { en: 'Unpublish', bg: 'Спри публикуването' },
-  'admin.content.guides.btn.edit': { en: 'Edit', bg: 'Редактирай' },
-  'admin.content.guides.btn.delete': { en: 'Delete', bg: 'Изтрий' },
-  'admin.content.guides.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
-  'admin.content.guides.btn.save': { en: 'Save', bg: 'Запази' },
-  'admin.content.guides.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.content.guides.dialog.title': { en: 'Delete guide', bg: 'Изтриване на ръководство' },
-  'admin.content.guides.dialog.body': { en: 'Delete "{{name}}"? This cannot be undone.', bg: 'Изтриване на "{{name}}"? Действието е необратимо.' },
-  'admin.content.guides.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.content.guides.dialog.confirm': { en: 'Delete', bg: 'Изтрий' },
-  'admin.content.guides.lbl.titleEn': { en: 'Title (EN)', bg: 'Заглавие (EN)' },
-  'admin.content.guides.lbl.titleBg': { en: 'Title (BG)', bg: 'Заглавие (BG)' },
-  'admin.content.guides.lbl.slug': { en: 'Slug', bg: 'Slug' },
-  'admin.content.guides.lbl.emoji': { en: 'Emoji', bg: 'Емоджи' },
-  'admin.content.guides.lbl.coverImage': { en: 'Cover Image', bg: 'Корица' },
-  'admin.content.guides.lbl.accessRoles': { en: 'Access Roles', bg: 'Роли за достъп' },
-  'admin.content.guides.lbl.bodyBlocks': { en: 'Body Blocks', bg: 'Блокове' },
-  'admin.content.guides.btn.copyUrl': { en: 'Copy URL', bg: 'Копирай URL' },
-  'admin.content.guides.btn.copied': { en: 'Copied!', bg: 'Копирано!' },
-  'admin.content.guides.btn.upload': { en: 'Upload', bg: 'Качи' },
-  'admin.content.guides.btn.uploading': { en: 'Uploading…', bg: 'Качване…' },
-  'admin.content.guides.placeholder.pasteUrl': { en: 'or paste URL…', bg: 'или постави URL…' },
-  'admin.content.guides.markdownHint': { en: '**bold**, _italic_ supported', bg: '**удебелен**, _курсив_ поддържани' },
-
-  // -- Content: Links --
-  'admin.content.links.lbl.labelEn': { en: 'Label (EN)', bg: 'Етикет (EN)' },
-  'admin.content.links.lbl.labelBg': { en: 'Label (BG)', bg: 'Етикет (BG)' },
-  'admin.content.links.lbl.url': { en: 'URL', bg: 'URL' },
-  'admin.content.links.placeholder.labelEn': { en: 'Label in English', bg: 'Етикет на английски' },
-  'admin.content.links.placeholder.labelBg': { en: 'Label in Bulgarian', bg: 'Етикет на български' },
-  'admin.content.links.btn.adding': { en: 'Adding…', bg: 'Добавяне…' },
-  'admin.content.links.btn.add': { en: 'Add link', bg: 'Добави връзка' },
-  'admin.content.links.empty': { en: 'No links yet. Add one above.', bg: 'Все още няма връзки. Добавете по-горе.' },
-  'admin.content.links.btn.edit': { en: 'Edit', bg: 'Редактирай' },
-  'admin.content.links.btn.delete': { en: 'Delete', bg: 'Изтрий' },
-  'admin.content.links.btn.toggleActive': { en: 'Active', bg: 'Активна' },
-  'admin.content.links.btn.toggleInactive': { en: 'Inactive', bg: 'Неактивна' },
-  'admin.content.links.drawer.editTitle': { en: 'Edit link', bg: 'Редактиране на връзка' },
-  'admin.content.links.drawer.editTitleNamed': { en: 'Edit: {{name}}', bg: 'Редактиране: {{name}}' },
-  'admin.content.links.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
-  'admin.content.links.btn.saveChanges': { en: 'Save changes', bg: 'Запази промените' },
-  'admin.content.links.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.content.links.dialog.title': { en: 'Delete link', bg: 'Изтриване на връзка' },
-  'admin.content.links.dialog.body': { en: 'Delete "{{name}}"? This cannot be undone.', bg: 'Изтриване на "{{name}}"? Действието е необратимо.' },
-  'admin.content.links.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.content.links.dialog.confirm': { en: 'Delete', bg: 'Изтрий' },
-
-  // -- Content: News --
-  'admin.content.news.btn.new': { en: '+ New post', bg: '+ Нова публикация' },
-  'admin.content.news.drawer.newTitle': { en: 'New post', bg: 'Нова публикация' },
-  'admin.content.news.drawer.editTitle': { en: 'Edit: {{name}}', bg: 'Редактиране: {{name}}' },
-  'admin.content.news.empty': { en: 'No news posts yet.', bg: 'Все още няма публикации.' },
-  'admin.content.news.btn.edit': { en: 'Edit', bg: 'Редактирай' },
-  'admin.content.news.btn.delete': { en: 'Delete', bg: 'Изтрий' },
-  'admin.content.news.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
-  'admin.content.news.btn.save': { en: 'Save', bg: 'Запази' },
-  'admin.content.news.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.content.news.btn.activate': { en: 'Activate', bg: 'Активирай' },
-  'admin.content.news.btn.deactivate': { en: 'Deactivate', bg: 'Деактивирай' },
-  'admin.content.news.btn.copyUrl': { en: 'Copy URL', bg: 'Копирай URL' },
-  'admin.content.news.btn.copied': { en: 'Copied!', bg: 'Копирано!' },
-  'admin.content.news.status.active': { en: 'Active', bg: 'Активна' },
-  'admin.content.news.status.inactive': { en: 'Inactive', bg: 'Неактивна' },
-  'admin.content.news.dialog.title': { en: 'Delete post', bg: 'Изтриване на публикация' },
-  'admin.content.news.dialog.body': { en: 'Delete "{{name}}"? This cannot be undone.', bg: 'Изтриване на "{{name}}"? Действието е необратимо.' },
-  'admin.content.news.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.content.news.dialog.confirm': { en: 'Delete', bg: 'Изтрий' },
-  'admin.content.news.lbl.titleEn': { en: 'Title (EN)', bg: 'Заглавие (EN)' },
-  'admin.content.news.lbl.titleBg': { en: 'Title (BG)', bg: 'Заглавие (BG)' },
-  'admin.content.news.lbl.slug': { en: 'Slug', bg: 'Slug' },
-  'admin.content.news.lbl.contentEn': { en: 'Content (EN)', bg: 'Съдържание (EN)' },
-  'admin.content.news.lbl.contentBg': { en: 'Content (BG)', bg: 'Съдържание (BG)' },
-  'admin.content.news.lbl.accessRoles': { en: 'Access Roles', bg: 'Роли за достъп' },
-  'admin.content.news.slug.hint': { en: "If left empty, this post won't show a link on the homepage.", bg: 'Ако е оставено празно, публикацията няма да показва връзка на началната страница.' },
-
-  // -- Calendar --
-  'admin.calendar.pageTitle': { en: 'Calendar Events', bg: 'Календарни събития' },
-  'admin.calendar.pageDesc': { en: 'Create, edit, and delete calendar events. Google Calendar sync also populates this table.', bg: 'Създавайте, редактирайте и изтривайте календарни събития. Синхронизацията с Google Calendar също попълва тази таблица.' },
-  'admin.calendar.btn.new': { en: '+ New event', bg: '+ Ново събитие' },
-  'admin.calendar.empty': { en: 'No events yet.', bg: 'Все още няма събития.' },
-  'admin.calendar.badge.google': { en: 'Google', bg: 'Google' },
-  'admin.calendar.badge.guestReg': { en: 'guest reg', bg: 'гост рег.' },
-  'admin.calendar.btn.edit': { en: 'Edit', bg: 'Редактирай' },
-  'admin.calendar.btn.delete': { en: 'Delete', bg: 'Изтрий' },
-  'admin.calendar.confirm.delete': { en: 'Delete "{{title}}"?', bg: 'Изтриване на "{{title}}"?' },
-  'admin.calendar.drawer.newTitle': { en: 'New event', bg: 'Ново събитие' },
-  'admin.calendar.drawer.editTitle': { en: 'Edit: {{title}}', bg: 'Редактиране: {{title}}' },
-  'admin.calendar.lbl.start': { en: 'Start', bg: 'Начало' },
-  'admin.calendar.lbl.end': { en: 'End', bg: 'Край' },
-  'admin.calendar.lbl.category': { en: 'Category', bg: 'Категория' },
-  'admin.calendar.lbl.type': { en: 'Type', bg: 'Тип' },
-  'admin.calendar.lbl.visibleTo': { en: 'Visible to', bg: 'Видимо за' },
-  'admin.calendar.lbl.meetingUrl': { en: 'Meeting URL', bg: 'URL за среща' },
-  'admin.calendar.lbl.allowGuestReg': { en: 'Allow guest registration', bg: 'Разреши регистрация на гости' },
-  'admin.calendar.lbl.availableRoles': { en: 'Available roles', bg: 'Налични роли' },
-  'admin.calendar.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
-  'admin.calendar.btn.saveChanges': { en: 'Save changes', bg: 'Запази промените' },
-  'admin.calendar.btn.createEvent': { en: 'Create event', bg: 'Създай събитие' },
-  'admin.calendar.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
-  'admin.calendar.allMonths': { en: 'All months', bg: 'Всички месеци' },
-  'admin.calendar.placeholder.title': { en: 'Title', bg: 'Заглавие' },
-  'admin.calendar.placeholder.description': { en: 'Description (optional)', bg: 'Описание (незадължително)' },
-  'admin.calendar.placeholder.meetingUrl': { en: 'https://zoom.us/j/…', bg: 'https://zoom.us/j/…' },
-  'admin.calendar.placeholder.roleTag': { en: 'Add role…', bg: 'Добави роля…' },
-
-} as const
+export { admin } from './admin/index'
