+++ lib/i18n/domains/admin/content.ts
@@ -1,0 +1,184 @@
+export const adminContent = {
+  // -- Settings --
+  'admin.settings.pageTitle': { en: 'System Settings', bg: 'Системни настройки' },
+  'admin.settings.pageDesc': { en: 'Manage global configurations and monitor automated systems', bg: 'Управлявайте глобалните конфигурации и наблюдавайте автоматизираните системи' },
+
+  // -- Settings: Email Log --
+  'admin.settings.emailLog.title': { en: 'Email Log', bg: 'Журнал на имейлите' },
+  'admin.settings.emailLog.desc': { en: 'Every outbound email attempt is recorded here. Use Retry on failed rows.', bg: 'Всеки опит за изпращане на имейл се записва тук. Използвайте Повторен опит за неуспешни редове.' },
+  'admin.settings.emailLog.col.template': { en: 'Template', bg: 'Шаблон' },
+  'admin.settings.emailLog.col.recipient': { en: 'Recipient', bg: 'Получател' },
+  'admin.settings.emailLog.col.sentAt': { en: 'Sent at', bg: 'Изпратено в' },
+  'admin.settings.emailLog.col.status': { en: 'Status', bg: 'Статус' },
+  'admin.settings.emailLog.filter.all': { en: 'All', bg: 'Всички' },
+  'admin.settings.emailLog.filter.sent': { en: 'sent', bg: 'изпратени' },
+  'admin.settings.emailLog.filter.failed': { en: 'failed', bg: 'неуспешни' },
+  'admin.settings.emailLog.filter.pending': { en: 'pending', bg: 'чакащи' },
+  'admin.settings.emailLog.allTemplates': { en: 'All templates', bg: 'Всички шаблони' },
+  'admin.settings.emailLog.empty': { en: 'No email log entries found.', bg: 'Няма намерени записи в журнала.' },
+  'admin.settings.emailLog.retryError': { en: 'Retry error:', bg: 'Грешка при повторен опит:' },
+  'admin.settings.emailLog.btn.retry': { en: 'Retry', bg: 'Повторен опит' },
+  'admin.settings.emailLog.btn.retrying': { en: 'Retrying…', bg: 'Повторен опит…' },
+  'admin.settings.emailLog.pagination.prev': { en: '← Prev', bg: '← Предишна' },
+  'admin.settings.emailLog.pagination.next': { en: 'Next →', bg: 'Следваща →' },
+
+  // -- Settings: Email Settings Panel --
+  'admin.settings.emailPanel.globalTitle': { en: 'System-wide Email', bg: 'Системни имейли' },
+  'admin.settings.emailPanel.globalDesc': { en: 'Enable or disable all automated outbound emails.', bg: 'Включете или изключете всички автоматизирани изходящи имейли.' },
+  'admin.settings.emailPanel.recipientTitle': { en: 'Admin Alert Recipient', bg: 'Получател на системни сигнали' },
+  'admin.settings.emailPanel.recipientDesc': { en: 'Receives bcc or system alerts if configured.', bg: 'Получава скрито копие или системни сигнали при конфигурация.' },
+  'admin.settings.emailPanel.btn.save': { en: 'Save', bg: 'Запази' },
+  'admin.settings.emailPanel.automatedTitle': { en: 'Automated Notifications', bg: 'Автоматизирани известия' },
+  'admin.settings.emailPanel.template.welcome': { en: 'Welcome Email', bg: 'Приветствен имейл' },
+  'admin.settings.emailPanel.template.payment_status': { en: 'Payment Status Updates', bg: 'Обновяване на статус на плащане' },
+  'admin.settings.emailPanel.template.doc_expiry': { en: 'Document Expiry Warnings', bg: 'Предупреждения за изтичащи документи' },
+  'admin.settings.emailPanel.template.abo_verification_result': { en: 'ABO Verification Results', bg: 'Резултати от СБА верификация' },
+  'admin.settings.emailPanel.template.trip_registration_status': { en: 'Trip Registration Updates', bg: 'Обновяване на регистрация за пътуване' },
+
+  // -- Content: page --
+  'admin.content.page.heading': { en: 'Content', bg: 'Съдържание' },
+  'admin.content.page.tab.news':    { en: 'News',         bg: 'Новини'        },
+  'admin.content.page.tab.links':   { en: 'Links',        bg: 'Връзки'        },
+  'admin.content.page.tab.guides':  { en: 'Guides',       bg: 'Ръководства'   },
+  'admin.content.page.tab.socials': { en: 'Social Posts', bg: 'Публикации'    },
+
+  // -- Content: Social --
+  'admin.content.social.btn.addPost': { en: '+ Add post', bg: '+ Добави публикация' },
+  'admin.content.social.drawer.newTitle': { en: 'New post', bg: 'Нова публикация' },
+  'admin.content.social.drawer.editTitle': { en: 'Edit: {{name}}', bg: 'Редактиране: {{name}}' },
+  'admin.content.social.placeholder.postUrl': { en: 'Post URL (required)', bg: 'URL на публикацията (задължително)' },
+  'admin.content.social.fetchingPreview': { en: 'Fetching preview…', bg: 'Зареждане на преглед…' },
+  'admin.content.social.previewUnavailable': { en: 'Preview unavailable for this platform — enter thumbnail URL manually', bg: 'Прегледът не е наличен за тази платформа — въведете URL на миниатюрата ръчно' },
+  'admin.content.social.placeholder.caption': { en: 'Caption — auto-extracted from post if left blank', bg: 'Надпис — автоматично извлечен от публикацията' },
+  'admin.content.social.placeholder.thumbnailUrl': { en: 'Thumbnail URL — auto-extracted from post if left blank', bg: 'URL на миниатюра — автоматично извлечен от публикацията' },
+  'admin.content.social.btn.uploading': { en: 'Uploading…', bg: 'Качване…' },
+  'admin.content.social.btn.uploadThumbnail': { en: 'Upload thumbnail', bg: 'Качи миниатюра' },
+  'admin.content.social.thumbnailStored': { en: 'Stored', bg: 'Съхранено' },
+  'admin.content.social.cdnUrlWarning': { en: 'This URL is an ephemeral CDN link and will expire. Upload a file instead.', bg: 'Този URL е временна CDN връзка и ще изтече. Качете файл вместо това.' },
+  'admin.content.social.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
+  'admin.content.social.btn.save': { en: 'Save', bg: 'Запази' },
+  'admin.content.social.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.content.social.badge.pinned': { en: 'Pinned', bg: 'Закачено' },
+  'admin.content.social.badge.active': { en: 'Active', bg: 'Активно' },
+  'admin.content.social.badge.hidden': { en: 'Hidden', bg: 'Скрито' },
+  'admin.content.social.btn.unpin': { en: 'Unpin', bg: 'Откачи' },
+  'admin.content.social.btn.pin': { en: 'Pin', bg: 'Закачи' },
+  'admin.content.social.btn.edit': { en: 'Edit', bg: 'Редактирай' },
+  'admin.content.social.btn.delete': { en: 'Delete', bg: 'Изтрий' },
+  'admin.content.social.empty': { en: 'No posts yet.', bg: 'Все още няма публикации.' },
+  'admin.content.social.dialog.title': { en: 'Delete post', bg: 'Изтриване на публикация' },
+  'admin.content.social.dialog.body': { en: 'Delete this post? This cannot be undone.', bg: 'Изтриване на тази публикация? Действието е необратимо.' },
+  'admin.content.social.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.content.social.dialog.confirm': { en: 'Delete', bg: 'Изтрий' },
+
+  // -- Content: Guides --
+  'admin.content.guides.btn.new': { en: '+ New Guide', bg: '+ Ново ръководство' },
+  'admin.content.guides.drawer.newTitle': { en: 'New Guide', bg: 'Ново ръководство' },
+  'admin.content.guides.drawer.editTitle': { en: 'Edit: {{name}}', bg: 'Редактиране: {{name}}' },
+  'admin.content.guides.empty': { en: 'No guides yet. Create the first one.', bg: 'Все още няма ръководства. Създайте първото.' },
+  'admin.content.guides.btn.publish': { en: 'Publish', bg: 'Публикувай' },
+  'admin.content.guides.btn.unpublish': { en: 'Unpublish', bg: 'Спри публикуването' },
+  'admin.content.guides.btn.edit': { en: 'Edit', bg: 'Редактирай' },
+  'admin.content.guides.btn.delete': { en: 'Delete', bg: 'Изтрий' },
+  'admin.content.guides.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
+  'admin.content.guides.btn.save': { en: 'Save', bg: 'Запази' },
+  'admin.content.guides.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.content.guides.dialog.title': { en: 'Delete guide', bg: 'Изтриване на ръководство' },
+  'admin.content.guides.dialog.body': { en: 'Delete "{{name}}"? This cannot be undone.', bg: 'Изтриване на "{{name}}"? Действието е необратимо.' },
+  'admin.content.guides.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.content.guides.dialog.confirm': { en: 'Delete', bg: 'Изтрий' },
+  'admin.content.guides.lbl.titleEn': { en: 'Title (EN)', bg: 'Заглавие (EN)' },
+  'admin.content.guides.lbl.titleBg': { en: 'Title (BG)', bg: 'Заглавие (BG)' },
+  'admin.content.guides.lbl.slug': { en: 'Slug', bg: 'Slug' },
+  'admin.content.guides.lbl.emoji': { en: 'Emoji', bg: 'Емоджи' },
+  'admin.content.guides.lbl.coverImage': { en: 'Cover Image', bg: 'Корица' },
+  'admin.content.guides.lbl.accessRoles': { en: 'Access Roles', bg: 'Роли за достъп' },
+  'admin.content.guides.lbl.bodyBlocks': { en: 'Body Blocks', bg: 'Блокове' },
+  'admin.content.guides.btn.copyUrl': { en: 'Copy URL', bg: 'Копирай URL' },
+  'admin.content.guides.btn.copied': { en: 'Copied!', bg: 'Копирано!' },
+  'admin.content.guides.btn.upload': { en: 'Upload', bg: 'Качи' },
+  'admin.content.guides.btn.uploading': { en: 'Uploading…', bg: 'Качване…' },
+  'admin.content.guides.placeholder.pasteUrl': { en: 'or paste URL…', bg: 'или постави URL…' },
+  'admin.content.guides.markdownHint': { en: '**bold**, _italic_ supported', bg: '**удебелен**, _курсив_ поддържани' },
+
+  // -- Content: Links --
+  'admin.content.links.lbl.labelEn': { en: 'Label (EN)', bg: 'Етикет (EN)' },
+  'admin.content.links.lbl.labelBg': { en: 'Label (BG)', bg: 'Етикет (BG)' },
+  'admin.content.links.lbl.url': { en: 'URL', bg: 'URL' },
+  'admin.content.links.placeholder.labelEn': { en: 'Label in English', bg: 'Етикет на английски' },
+  'admin.content.links.placeholder.labelBg': { en: 'Label in Bulgarian', bg: 'Етикет на български' },
+  'admin.content.links.btn.adding': { en: 'Adding…', bg: 'Добавяне…' },
+  'admin.content.links.btn.add': { en: 'Add link', bg: 'Добави връзка' },
+  'admin.content.links.empty': { en: 'No links yet. Add one above.', bg: 'Все още няма връзки. Добавете по-горе.' },
+  'admin.content.links.btn.edit': { en: 'Edit', bg: 'Редактирай' },
+  'admin.content.links.btn.delete': { en: 'Delete', bg: 'Изтрий' },
+  'admin.content.links.btn.toggleActive': { en: 'Active', bg: 'Активна' },
+  'admin.content.links.btn.toggleInactive': { en: 'Inactive', bg: 'Неактивна' },
+  'admin.content.links.drawer.editTitle': { en: 'Edit link', bg: 'Редактиране на връзка' },
+  'admin.content.links.drawer.editTitleNamed': { en: 'Edit: {{name}}', bg: 'Редактиране: {{name}}' },
+  'admin.content.links.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
+  'admin.content.links.btn.saveChanges': { en: 'Save changes', bg: 'Запази промените' },
+  'admin.content.links.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.content.links.dialog.title': { en: 'Delete link', bg: 'Изтриване на връзка' },
+  'admin.content.links.dialog.body': { en: 'Delete "{{name}}"? This cannot be undone.', bg: 'Изтриване на "{{name}}"? Действието е необратимо.' },
+  'admin.content.links.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.content.links.dialog.confirm': { en: 'Delete', bg: 'Изтрий' },
+
+  // -- Content: News --
+  'admin.content.news.btn.new': { en: '+ New post', bg: '+ Нова публикация' },
+  'admin.content.news.drawer.newTitle': { en: 'New post', bg: 'Нова публикация' },
+  'admin.content.news.drawer.editTitle': { en: 'Edit: {{name}}', bg: 'Редактиране: {{name}}' },
+  'admin.content.news.empty': { en: 'No news posts yet.', bg: 'Все още няма публикации.' },
+  'admin.content.news.btn.edit': { en: 'Edit', bg: 'Редактирай' },
+  'admin.content.news.btn.delete': { en: 'Delete', bg: 'Изтрий' },
+  'admin.content.news.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
+  'admin.content.news.btn.save': { en: 'Save', bg: 'Запази' },
+  'admin.content.news.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.content.news.btn.activate': { en: 'Activate', bg: 'Активирай' },
+  'admin.content.news.btn.deactivate': { en: 'Deactivate', bg: 'Деактивирай' },
+  'admin.content.news.btn.copyUrl': { en: 'Copy URL', bg: 'Копирай URL' },
+  'admin.content.news.btn.copied': { en: 'Copied!', bg: 'Копирано!' },
+  'admin.content.news.status.active': { en: 'Active', bg: 'Активна' },
+  'admin.content.news.status.inactive': { en: 'Inactive', bg: 'Неактивна' },
+  'admin.content.news.dialog.title': { en: 'Delete post', bg: 'Изтриване на публикация' },
+  'admin.content.news.dialog.body': { en: 'Delete "{{name}}"? This cannot be undone.', bg: 'Изтриване на "{{name}}"? Действието е необратимо.' },
+  'admin.content.news.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.content.news.dialog.confirm': { en: 'Delete', bg: 'Изтрий' },
+  'admin.content.news.lbl.titleEn': { en: 'Title (EN)', bg: 'Заглавие (EN)' },
+  'admin.content.news.lbl.titleBg': { en: 'Title (BG)', bg: 'Заглавие (BG)' },
+  'admin.content.news.lbl.slug': { en: 'Slug', bg: 'Slug' },
+  'admin.content.news.lbl.contentEn': { en: 'Content (EN)', bg: 'Съдържание (EN)' },
+  'admin.content.news.lbl.contentBg': { en: 'Content (BG)', bg: 'Съдържание (BG)' },
+  'admin.content.news.lbl.accessRoles': { en: 'Access Roles', bg: 'Роли за достъп' },
+  'admin.content.news.slug.hint': { en: "If left empty, this post won't show a link on the homepage.", bg: 'Ако е оставено празно, публикацията няма да показва връзка на началната страница.' },
+
+  // -- Calendar --
+  'admin.calendar.pageTitle': { en: 'Calendar Events', bg: 'Календарни събития' },
+  'admin.calendar.pageDesc': { en: 'Create, edit, and delete calendar events. Google Calendar sync also populates this table.', bg: 'Създавайте, редактирайте и изтривайте календарни събития. Синхронизацията с Google Calendar също попълва тази таблица.' },
+  'admin.calendar.btn.new': { en: '+ New event', bg: '+ Ново събитие' },
+  'admin.calendar.empty': { en: 'No events yet.', bg: 'Все още няма събития.' },
+  'admin.calendar.badge.google': { en: 'Google', bg: 'Google' },
+  'admin.calendar.badge.guestReg': { en: 'guest reg', bg: 'гост рег.' },
+  'admin.calendar.btn.edit': { en: 'Edit', bg: 'Редактирай' },
+  'admin.calendar.btn.delete': { en: 'Delete', bg: 'Изтрий' },
+  'admin.calendar.confirm.delete': { en: 'Delete "{{title}}"?', bg: 'Изтриване на "{{title}}"?' },
+  'admin.calendar.drawer.newTitle': { en: 'New event', bg: 'Ново събитие' },
+  'admin.calendar.drawer.editTitle': { en: 'Edit: {{title}}', bg: 'Редактиране: {{title}}' },
+  'admin.calendar.lbl.start': { en: 'Start', bg: 'Начало' },
+  'admin.calendar.lbl.end': { en: 'End', bg: 'Край' },
+  'admin.calendar.lbl.category': { en: 'Category', bg: 'Категория' },
+  'admin.calendar.lbl.type': { en: 'Type', bg: 'Тип' },
+  'admin.calendar.lbl.visibleTo': { en: 'Visible to', bg: 'Видимо за' },
+  'admin.calendar.lbl.meetingUrl': { en: 'Meeting URL', bg: 'URL за среща' },
+  'admin.calendar.lbl.allowGuestReg': { en: 'Allow guest registration', bg: 'Разреши регистрация на гости' },
+  'admin.calendar.lbl.availableRoles': { en: 'Available roles', bg: 'Налични роли' },
+  'admin.calendar.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
+  'admin.calendar.btn.saveChanges': { en: 'Save changes', bg: 'Запази промените' },
+  'admin.calendar.btn.createEvent': { en: 'Create event', bg: 'Създай събитие' },
+  'admin.calendar.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.calendar.allMonths': { en: 'All months', bg: 'Всички месеци' },
+  'admin.calendar.placeholder.title': { en: 'Title', bg: 'Заглавие' },
+  'admin.calendar.placeholder.description': { en: 'Description (optional)', bg: 'Описание (незадължително)' },
+  'admin.calendar.placeholder.meetingUrl': { en: 'https://zoom.us/j/…', bg: 'https://zoom.us/j/…' },
+  'admin.calendar.placeholder.roleTag': { en: 'Add role…', bg: 'Добави роля…' },
+} as const
