+++ lib/i18n/domains/admin/index.ts
@@ -1,0 +1,9 @@
+import { adminCore } from './core'
+import { adminOperations } from './operations'
+import { adminContent } from './content'
+
+export const admin = {
+  ...adminCore,
+  ...adminOperations,
+  ...adminContent,
+} as const
