+++ lib/i18n/domains/admin/operations.ts
@@ -1,0 +1,137 @@
+export const adminOperations = {
+  // -- Approval Hub --
+  'admin.approval.verify.standardTitle': { en: 'Standard ABO requests — {{count}}', bg: 'Стандартни СБА заявки — {{count}}' },
+  'admin.approval.verify.noStandard': { en: 'No pending standard requests.', bg: 'Няма чакащи стандартни заявки.' },
+  'admin.approval.verify.manualTitle': { en: 'Manual requests — {{count}}', bg: 'Ръчни заявки — {{count}}' },
+  'admin.approval.verify.noManual': { en: 'No pending manual requests.', bg: 'Няма чакащи ръчни заявки.' },
+
+  'admin.approval.verify.noAboBadge': { en: 'No ABO', bg: 'Без СБА' },
+  'admin.approval.verify.awaitingPosTitle': { en: 'Awaiting LOS positioning — {{count}}', bg: 'Очаква позициониране в LOS — {{count}}' },
+  'admin.approval.verify.approvedBadge': { en: 'approved', bg: 'одобрен' },
+
+  'admin.approval.verify.directTitle': { en: 'Direct verify', bg: 'Директна верификация' },
+  'admin.approval.verify.directDesc': { en: 'Directly promote a guest to member without a prior submission.', bg: 'Пряко повишаване на гост до член без предварителна заявка.' },
+  'admin.approval.verify.lbl.guest': { en: 'Guest', bg: 'Гост' },
+  'admin.approval.verify.lbl.uplineAbo': { en: 'Upline ABO number', bg: 'СБА номер на горна линия' },
+  'admin.approval.verify.opt.selectGuest': { en: 'Select a guest…', bg: 'Избери гост…' },
+  'admin.approval.verify.placeholder.upline': { en: 'e.g. 7010970187', bg: 'напр. 7010970187' },
+  'admin.approval.verify.btn.approve': { en: 'Approve', bg: 'Одобри' },
+  'admin.approval.verify.btn.deny': { en: 'Deny', bg: 'Откажи' },
+  'admin.approval.verify.btn.verifyMember': { en: 'Verify member', bg: 'Верифицирай член' },
+  'admin.approval.verify.btn.verifying': { en: 'Verifying…', bg: 'Верифициране…' },
+  'admin.approval.verify.btn.verified': { en: 'Verified ✓', bg: 'Верифициран ✓' },
+  'admin.approval.verify.noGuests': { en: 'No guests without a pending request.', bg: 'Няма гости без чакаща заявка.' },
+
+  'admin.approval.trips.resolvedCollapsible': { en: 'Resolved — {{count}}', bg: 'Разрешени — {{count}}' },
+  'admin.approval.trips.btn.allTrips': { en: 'All trips', bg: 'Всички пътувания' },
+  'admin.approval.trips.pendingTitle': { en: 'Pending — {{count}}', bg: 'Чакащи — {{count}}' },
+  'admin.approval.trips.noPending': { en: 'No pending registrations.', bg: 'Няма чакащи регистрации.' },
+
+  'admin.approval.events.btn.allEvents': { en: 'All events', bg: 'Всички събития' },
+  'admin.approval.events.noPending': { en: 'No pending role requests.', bg: 'Няма чакащи заявки за роли.' },
+
+  // -- Operations: Trips --
+  'admin.operations.trips.btn.new': { en: '+ New trip', bg: '+ Ново пътуване' },
+  'admin.operations.trips.empty': { en: 'No trips yet.', bg: 'Все още няма пътувания.' },
+  'admin.operations.trips.milestones': { en: '{{count}} milestones', bg: '{{count}} вноски' },
+  'admin.operations.trips.btn.edit': { en: 'Edit', bg: 'Редактирай' },
+  'admin.operations.trips.btn.delete': { en: 'Delete', bg: 'Изтрий' },
+  'admin.operations.trips.title.edit': { en: 'Edit: {{title}}', bg: 'Редактиране: {{title}}' },
+  'admin.operations.trips.dialog.title': { en: 'Delete trip', bg: 'Изтриване на пътуване' },
+  'admin.operations.trips.dialog.body': { en: 'This cannot be undone.', bg: 'Това действие е необратимо.' },
+  'admin.operations.trips.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.operations.trips.dialog.confirm': { en: 'Delete', bg: 'Изтрий' },
+
+  // -- Operations: TripForm --
+  'admin.operations.form.documents': { en: 'Trip documents', bg: 'Документи за пътуването' },
+  'admin.operations.form.loading': { en: 'Loading…', bg: 'Зареждане…' },
+  'admin.operations.form.noDocuments': { en: 'No documents uploaded yet.', bg: 'Все още няма качени документи.' },
+  'admin.operations.form.btn.uploading': { en: 'Uploading…', bg: 'Качване…' },
+  'admin.operations.form.btn.upload': { en: '+ Upload file', bg: '+ Качи файл' },
+  'admin.operations.form.uploadHint': { en: 'PDF or image, max 10 MB', bg: 'PDF или изображение, макс. 10 МБ' },
+  'admin.operations.form.btn.remove': { en: 'Remove', bg: 'Премахни' },
+  'admin.operations.form.dialog.title': { en: 'Remove document', bg: 'Премахване на документ' },
+  'admin.operations.form.dialog.removeConfirm': { en: 'Remove "{{name}}"?', bg: 'Премахване на "{{name}}"?' },
+  'admin.operations.form.dialog.body': { en: 'This cannot be undone.', bg: 'Това действие е необратимо.' },
+  'admin.operations.form.dialog.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.operations.form.dialog.confirm': { en: 'Remove', bg: 'Премахни' },
+  'admin.operations.form.lbl.description': { en: 'Description', bg: 'Описание' },
+  'admin.operations.form.lbl.startDate': { en: 'Start date', bg: 'Начална дата' },
+  'admin.operations.form.lbl.endDate': { en: 'End date', bg: 'Крайна дата' },
+  'admin.operations.form.lbl.cost': { en: 'Cost (EUR)', bg: 'Цена (EUR)' },
+  'admin.operations.form.lbl.milestones': { en: 'Payment milestones', bg: 'Вноски' },
+  'admin.operations.form.placeholder.label': { en: 'Label', bg: 'Наименование' },
+  'admin.operations.form.placeholder.amount': { en: 'Amount', bg: 'Сума' },
+  'admin.operations.form.btn.addMilestone': { en: '+ Add', bg: '+ Добави' },
+  'admin.operations.form.lbl.location': { en: 'Location', bg: 'Местоположение' },
+  'admin.operations.form.lbl.tripType': { en: 'Trip type', bg: 'Тип пътуване' },
+  'admin.operations.form.lbl.accommodation': { en: 'Accommodation', bg: 'Настаняване' },
+  'admin.operations.form.lbl.inclusions': { en: 'Inclusions (comma-sep)', bg: 'Включено (разделено със запетая)' },
+  'admin.operations.form.lbl.visibleTo': { en: 'Visible to', bg: 'Видимо за' },
+  'admin.operations.form.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
+  'admin.operations.form.btn.saveChanges': { en: 'Save changes', bg: 'Запази промените' },
+  'admin.operations.form.btn.createTrip': { en: 'Create trip', bg: 'Създай пътуване' },
+  'admin.operations.form.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
+
+  // -- Operations: Payments --
+  'admin.operations.payments.btn.log': { en: 'Log Payment', bg: 'Регистрирай плащане' },
+  'admin.operations.payments.pendingTitle': { en: 'Pending submissions ({{count}})', bg: 'Чакащи подавания ({{count}})' },
+  'admin.operations.payments.viewProof': { en: 'View proof ↗', bg: 'Виж доказателство ↗' },
+  'admin.operations.payments.badge.pending': { en: 'pending', bg: 'чакащо' },
+  'admin.operations.payments.filter.all': { en: 'All', bg: 'Всички' },
+  'admin.operations.payments.filter.pending': { en: 'Pending', bg: 'Чакащи' },
+  'admin.operations.payments.filter.approved': { en: 'Approved', bg: 'Одобрени' },
+  'admin.operations.payments.filter.rejected': { en: 'Rejected', bg: 'Отказани' },
+  'admin.operations.payments.empty': { en: 'No payments found.', bg: 'Няма намерени плащания.' },
+  'admin.operations.payments.drawer.title': { en: 'Log Payment', bg: 'Регистрирай плащане' },
+  'admin.operations.payments.lbl.entity': { en: 'Entity *', bg: 'Обект *' },
+  'admin.operations.payments.placeholder.entity': { en: 'Select entity…', bg: 'Избери обект…' },
+  'admin.operations.payments.group.trips': { en: 'Trips', bg: 'Пътувания' },
+  'admin.operations.payments.group.items': { en: 'Items', bg: 'Артикули' },
+  'admin.operations.payments.lbl.member': { en: 'Member *', bg: 'Член *' },
+  'admin.operations.payments.placeholder.member': { en: 'Select member…', bg: 'Избери член…' },
+  'admin.operations.payments.lbl.amount': { en: 'Amount *', bg: 'Сума *' },
+  'admin.operations.payments.lbl.currency': { en: 'Currency', bg: 'Валута' },
+  'admin.operations.payments.lbl.date': { en: 'Date', bg: 'Дата' },
+  'admin.operations.payments.lbl.method': { en: 'Payment method', bg: 'Начин на плащане' },
+  'admin.operations.payments.placeholder.method': { en: 'e.g. bank transfer, cash', bg: 'напр. банков превод, в брой' },
+  'admin.operations.payments.lbl.status': { en: 'Status', bg: 'Статус' },
+  'admin.operations.payments.lbl.note': { en: 'Note', bg: 'Бележка' },
+  'admin.operations.payments.lbl.noteOptional': { en: '(optional)', bg: '(незадължително)' },
+  'admin.operations.payments.placeholder.note': { en: 'e.g. Cash payment, ref #123', bg: 'напр. Плащане в брой, реф. #123' },
+  'admin.operations.payments.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
+  'admin.operations.payments.btn.log2': { en: 'Log payment', bg: 'Регистрирай плащане' },
+  'admin.operations.payments.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.operations.payments.btn.approve': { en: 'Approve', bg: 'Одобри' },
+  'admin.operations.payments.btn.deny': { en: 'Deny', bg: 'Откажи' },
+  'admin.operations.payments.placeholder.adminNote': { en: 'Admin note (optional)', bg: 'Бележка на администратор (незадължително)' },
+  'admin.operations.payments.status.approved': { en: 'Approved', bg: 'Одобрено' },
+  'admin.operations.payments.status.pending': { en: 'Pending', bg: 'Чакащо' },
+
+  // -- Operations: Items --
+  'admin.operations.items.btn.new': { en: '+ New item', bg: '+ Нов артикул' },
+  'admin.operations.items.empty': { en: 'No payable items yet.', bg: 'Все още няма артикули.' },
+  'admin.operations.items.btn.edit': { en: 'Edit', bg: 'Редактирай' },
+  'admin.operations.items.btn.deactivate': { en: 'Deactivate', bg: 'Деактивирай' },
+  'admin.operations.items.badge.active': { en: 'active', bg: 'активен' },
+  'admin.operations.items.badge.inactive': { en: 'inactive', bg: 'неактивен' },
+  'admin.operations.items.title.edit': { en: 'Edit: {{title}}', bg: 'Редактиране: {{title}}' },
+  'admin.operations.items.lbl.title': { en: 'Title *', bg: 'Наименование *' },
+  'admin.operations.items.lbl.type': { en: 'Type *', bg: 'Тип *' },
+  'admin.operations.items.lbl.description': { en: 'Description', bg: 'Описание' },
+  'admin.operations.items.lbl.amount': { en: 'Amount *', bg: 'Сума *' },
+  'admin.operations.items.lbl.currency': { en: 'Currency', bg: 'Валута' },
+  'admin.operations.items.lbl.linkedTrip': { en: 'Linked trip', bg: 'Свързано пътуване' },
+  'admin.operations.items.lbl.noTrip': { en: 'No trip', bg: 'Без пътуване' },
+  'admin.operations.items.toggle.active': { en: 'Active', bg: 'Активен' },
+  'admin.operations.items.toggle.inactive': { en: 'Inactive', bg: 'Неактивен' },
+  'admin.operations.items.btn.saving': { en: 'Saving…', bg: 'Запазване…' },
+  'admin.operations.items.btn.saveChanges': { en: 'Save changes', bg: 'Запази промените' },
+  'admin.operations.items.btn.create': { en: 'Create item', bg: 'Създай артикул' },
+  'admin.operations.items.btn.cancel': { en: 'Cancel', bg: 'Откажи' },
+  'admin.operations.items.type.merchandise': { en: 'merchandise', bg: 'стока' },
+  'admin.operations.items.type.ticket': { en: 'ticket', bg: 'билет' },
+  'admin.operations.items.type.food': { en: 'food', bg: 'храна' },
+  'admin.operations.items.type.book': { en: 'book', bg: 'книга' },
+  'admin.operations.items.type.other': { en: 'other', bg: 'друго' },
+} as const
